 requirements.txt
streamlit
numpy
Pillow
opencv-python
torch
torchvision
pytesseract
transformers
matplotlib
scikit-learn
